claude-hub/requirements.txt
# Claude Hub - Python Dependencies
# FastAPI Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Supabase Client
supabase>=2.3.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0

# Date/Time
python-dateutil>=2.8.0

# Templates (HTML Rendering)
jinja2>=3.1.0

# Static Files
aiofiles>=23.0.0
